Remove my injected layers
